task 1 + 2

+++ scrapper/src/main/java/ru/khanin/dmitrii/DTO/LinkResponse.java
@@ -2,4 +2,4 @@ package ru.khanin.dmitrii.DTO;
 
 import java.net.URI;
 
-public record LinkResponse(int id, URI url) {}
+public record LinkResponse(long id, URI url) {}
